+++ lib/glimmer/libui/control_proxy.rb
@@ -55,12 +55,12 @@ module Glimmer
           control_proxies.find {|c| c.is_a?(Glimmer::LibUI::WindowProxy)}
         end
         
-        def integer_to_boolean(int)
-          int.nil? ? nil : int == 1
+        def integer_to_boolean(int, allow_nil: true)
+          int.nil? ? (allow_nil ? nil : false) : int == 1
         end
         
-        def boolean_to_integer(bool)
-          bool.nil? ? nil : (bool ? 1 : 0)
+        def boolean_to_integer(bool, allow_nil: true)
+          bool.nil? ? (allow_nil ? nil : 0) : (bool ? 1 : 0)
         end
         
         def menu_proxies